wip: calculated incompatibilities
A lot of previously developed but only now used scaffolding was also
committed. This commit starts the actual work on the backtrack solver,
by computing the pairs of indices on the grid that need to be checked,
trying to be clever about it, and then constructing the list of
incompatible digits for each cell.