Messy commit: amend + annealing working(?)
Apparently there were some workspace-refactor related changes that I'd
missed commiting before. This is why you have some renames of (...).rs >
src/(...).rs. Otherwise, the work in this commit relates to getting the
annealing to working state.

It seems to indeed be working, with the simple.sudoku and
linear.schedule. The annealing with an initialization hint has NOT yet
been tested, neither have bigger commits (that are harder to debug,
especially now that a schedule can be responsible for a failure to get
to the ground state.