Projection solver working!(???)
This one was awful, and contrary to expectations it seems to yield
really bad results. Nonetheless, it was able to get to a solution in
example.sudoku, which is bound not to be coincidence. (However, for
comparison, it seems to have an awfully hard time with simple.sudoku).

The biggest pain point in the implementation right now is the way that
some digits are excluded; a big HashSet is created at the start with
each invalid digit for each cell. That's a big HashSet. However, a
number of previous approaches yielded wrong results, and I cannot tell
if there is an efficient way to encode this information.

In any case, even though the code by Chi and Lange is a mess, I think
that they 1. are using some sort of matricial representation --- it is
not specified in the paper --- and 2. this may end up being equivalent
to what I'm doing in terms of memory.

I do have a sneaking suspicion that no, the HashMap is redundant, but
it's working.