Refactored projection solver
Not sure this is any better, but at least it's clear what's happening
now, rather than a brute force approach of building a "forbiddens"
dictionary.

What I'm doing here is pre-computing the actual (to-be) simplexes that I
want to project, by looking at each potential row, column, subgrid, and
cell simplex, and removing the digits that are incompatible. One tricky
thing in doing this (and the reason why it currently requires so many
checks, if I'm not missing anything), is that there can be a digit that
is not of row or column (r, c), but preventing the existance of a digit
at a particular position of that row or column. Take, for example

   X _ _ _
   _ _ 4 _
   _ _ _ _
   _ _ _ _

Even though 4 is neither in the row nor in the column of X, it excludes
P[0, 2, 3] from appearing in the row simplex of X.

I think that now this is much closer to the (?) matrix approach of Chi
and Lange that I mentioned in the previous commit.

Again, it seems to have trouble with the simple sudoku, but converges
example.sudoku in 10_000 iterations.