Added benchmarking functionality and scripts
Also squashed a very bad bug on backtrack: it was enumerating the trees
in the wrong order of size! This effectively maximized the running time.

This is because I'm starting to do the benchmarking work, and finding
that moderately hard puzzles (such as the top dataset) absolutely
destroy my implementations.