Amend: filtering before enumeration broke backtrack
Now randomizing the order in which digits are tried for backtracking, in
order to try to avoid pathological cases in which the last digit is
always the feasible one.

Also, this has confirmed that what the paper considers to be "hard"
puzzles are, in fact, extremely easy. In particular, ~0.5s to solve the
hard puzzle provided by the authors vs. 10m+ unsolved for a single
top1465 puzzle.